css file path was still incorrect
Sorry there are like a million commits, here is a summary of my changes in this repository:
I had a lot of problems getting github to find my files and I had to keep adjusting file paths. At a point it worked but I wanted to try and see if I could put the index.html file in an html folder to make it more organized. I found out that even though that works locally on my pc, github wasn't able to find the index like that. I tried to put a dummy index.html in the root to redirect to the html in the folder, and it seemed to have worked but github still wasn't finding the index file. Eventually I deleted that and changed it back to the way it was before, but there were some random issues merging, so I removed the html file entirely and put it back in. Finally changed all the file paths to the way they were before to link the style sheet and the image properly. 

Throughout all these commits, nothing had changed in my assignment except for file paths, my assignment was completed a long time ago. In fact, the actual assignment was way easier than getting github to work properly. I am hoping this last change will finally get everything to display properly but if it doesn't, then just know i'm only making changes for the sake of getting github to work, not because I'm trying to add things to my assignment after submission. 

Thank you for reading.